agent/requirements.txt
Pillow==10.1.0
argparse
json5==0.9.14